+++ melting-java/src/melting/approximativeMethods/Owen69.java
@@ -26,7 +26,7 @@ public class Owen69 extends ApproximativeMode{
 	private static String temperatureEquation = "Tm = 87.16 + 0.345 * percentGC + log10(Na) * (20.17 - 0.066 * percentGC)";
 	
 	public ThermoResult CalculateThermodynamics() {
-		int percentGC = this.environment.getSequences().calculatePercentGC();
+		double percentGC = this.environment.getSequences().calculatePercentGC();
 		double Tm = super.CalculateThermodynamics().getTm(); 
 		Tm = 87.16 + 0.345 * percentGC + Math.log10(this.environment.getNa()) * (20.17 - 0.066 * percentGC);
